requirements.txt
requests==2.32.3
openai>=1.0.0
anthropic>=0.40.0
google-generativeai>=0.8.0
notion-client>=2.2.1
